Issue #474 Change return value of spaceBetweenConnectionPoints to int
There was no reasons to use a float there, since we are dealing with a
space in pixels.

+++ src/org/jetuml/rendering/edges/NodeIndex.java
@@ -76,7 +76,7 @@ public enum NodeIndex
 	 * @return the spacing in between connection points on pNodeFace. 
 	 * @pre pNodeFace != null && pAttachmentSide != null
 	 */
-	private static float spaceBetweenConnectionPoints(Line pNodeFace, Side pAttachmentSide)
+	private static int spaceBetweenConnectionPoints(Line pNodeFace, Side pAttachmentSide)
 	{
 		assert pNodeFace != null && pAttachmentSide != null;
 		if(pAttachmentSide.isHorizontal())
